requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
pydantic==2.5.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
kafka-python==2.0.2
httpx==0.25.2
websockets==12.0
starlette 